requirements.txt
django-tinymce
django-bootstrap3